move ThreeTierChoiceCards from epic to shared directory
